third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-size-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['background-size'] = "1px" should set the property value
  assert_equals: serialization should be canonical expected "1px auto" but got "1px"
[FAIL] e.style['background-size'] = "1px auto" should set the property value
  assert_equals: serialization should be canonical expected "1px auto" but got "1px"
Harness: the test ran to completion.
